Changed get Player.getJerseynumber to properly return byte (was short). Added PlayerTests (commented away the old ones)

+++ src/com/louislife/model/Player.java
@@ -80,7 +80,7 @@ public class Player {
 		this.surname = surname;
 	}
 
-	public short getJerseyNumber() {
+	public byte getJerseyNumber() {
 		return number;
 	}
